requirements.txt
absl-py~=0.7.1
appdirs==1.4.3
asn1crypto==0.24.0
astor==0.8.0
astroid==2.1.0
asttokens==1.1.13
beautifulsoup4==4.7.1
blinker~=1.4
certifi==2018.8.24
chardet==3.0.4
Click==7.0
colorama==0.3.7
colorzero==1.1
cookies==2.2.1
cryptography==2.6.1
cycler==0.10.0
decorator==4.3.0
docutils==0.14
entrypoints==0.3
Flask==1.0.2
flask-bower~=1.3.0
gast~=0.2.2
google-pasta==0.1.7
grpcio==1.22.0
guizero==0.6.0
h5py==2.9.0
html5lib==1.0.1
idna==2.6
ipykernel==4.9.0
ipython==5.8.0
ipython-genutils==0.2.0
isort==4.3.4
itsdangerous~=0.24
jedi==0.13.2
Jinja2>=2.10.1
jupyter-client==5.2.3
jupyter-core==4.4.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
keyring==17.1.1
keyrings.alt==3.1.1
kiwisolver==1.0.1
lazy-object-proxy~=1.3.1
logilab-common~=1.4.2
lxml==4.3.2
Markdown==3.1.1
MarkupSafe==1.1.0
matplotlib==3.0.2
mccabe==0.6.1
mote==0.0.4
mypy==0.670
mypy-extensions==0.4.1
nudatus~=0.0.4
numpy==1.16.2
oauthlib==2.1.0
olefile~=0.46
parso==0.3.1
pexpect==4.6.0
pgzero==1.2
pickleshare==0.7.5
Pillow==5.4.1
prompt-toolkit==1.0.15
protobuf==3.9.1
psutil~=5.5.1
pycairo~=1.16.2
pycodestyle~=2.4.0
pycrypto~=2.6.1
pyflakes==2.0.0
Pygments==2.3.1
pyinotify~=0.9.6
PyGObject~=3.30.4
PyJWT==1.7.0
pylint==2.2.2
pyOpenSSL==19.0.0
pyparsing==2.2.0
python-dateutil==2.7.3
pyxdg~=0.26
PyYAML~=5.1.2
pyzmq==17.1.2
requests==2.21.0
requests-oauthlib==1.0.0
responses==0.9.0
SecretStorage~=2.3.1
setuptools~=41.0.0
semver~=2.0.1
simplegeneric~=0.8.1
simplejson~=3.16.0
six==1.12.0
soupsieve==1.8
svgwrite==1.3.1
tensorboard==1.14.0
tensorflow==1.14.0
tensorflow-estimator==1.14.0
termcolor~=1.1.0
traitlets==4.3.2
typed-ast==1.3.1
urllib3==1.24.2
virtualenv==15.1.0
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.15.3
wrapt~=1.11.2
